+++ packages/editor/editor-confluence-transformer/__tests-karma__/browser/_schema-builder.ts
@@ -126,7 +126,7 @@ export const bodiedExtension = (
     extensionType: string;
     parameters?: object;
   },
-  content,
+  content: Node | Array<Node>,
 ) => schema.nodes.bodiedExtension.createChecked(attrs, content);
 export const emoji = (attrs: {
   id?: string;