+++ packages/editor/editor-core/src/__tests__/unit/plugins/insert-block/ui/ToolbarInsertBlock.tsx
@@ -38,7 +38,7 @@ import {
 } from '../../../../../plugins/hyperlink/pm-plugins/main';
 import {
   INPUT_METHOD,
-  AnalyticsEventPayload,
+  DispatchAnalyticsEvent,
 } from '../../../../../plugins/analytics';
 import tablesPlugin from '../../../../../plugins/table';
 import { AnalyticsHandler } from '../../../../../analytics';
@@ -108,9 +108,7 @@ describe('@atlaskit/editor-core/ui/ToolbarInsertBlock', () => {
   let toolbarOption: ReactWrapper;
   let analyticsHandlerSpy: jest.Mock<AnalyticsHandler>;
   let createAnalyticsEvent: CreateUIAnalyticsEventSignature;
-  let dispatchAnalyticsSpy: jest.SpyInstance<
-    (payload: AnalyticsEventPayload) => void
-  >;
+  let dispatchAnalyticsSpy: jest.SpyInstance<DispatchAnalyticsEvent>;
   let dispatchSpy: jest.SpyInstance;
 
   const editor = (doc: any, editorPlugins?: any[]) => {