+++ packages/editor/renderer/src/ui/MediaCard.tsx
@@ -173,10 +173,37 @@ export class MediaCardInternal extends Component<MediaCardProps, State> {
     );
   }
 
+  /**
+   * We want to call provided `eventHandlers.media.onClick` when it's provided,
+   * but we also don't want to call it when it's a video and inline video player is enabled.
+   * This is due to consumers normally process this onClick call by opening media viewer and
+   * we don't want that to happened described above text.
+   */
+  private getOnCardClickCallback = (isInlinePlayer: boolean) => {
+    const { eventHandlers } = this.props;
+    if (eventHandlers && eventHandlers.media && eventHandlers.media.onClick) {
+      return ((result, analyticsEvent) => {
+        const isVideo =
+          result.mediaItemDetails &&
+          result.mediaItemDetails.mediaType === 'video';
+        const isVideoWithInlinePlayer = isInlinePlayer && isVideo;
+        if (
+          !isVideoWithInlinePlayer &&
+          eventHandlers &&
+          eventHandlers.media &&
+          eventHandlers.media.onClick
+        ) {
+          eventHandlers.media.onClick(result, analyticsEvent);
+        }
+      }) as CardOnClickCallback;
+    }
+
+    return undefined;
+  };
+
   render() {
     const { context } = this.state;
     const {
-      eventHandlers,
       id,
       type,
       collection,
@@ -190,8 +217,10 @@ export class MediaCardInternal extends Component<MediaCardProps, State> {
     const isMobile = rendererAppearance === 'mobile';
     const shouldPlayInline =
       useInlinePlayer !== undefined ? useInlinePlayer : true;
-    const onCardClick =
-      eventHandlers && eventHandlers.media && eventHandlers.media.onClick;
+    const isInlinePlayer = isMobile ? false : shouldPlayInline;
+
+    const onCardClick = this.getOnCardClickCallback(isInlinePlayer);
+
     const shouldOpenMediaViewer = !isMobile && !onCardClick;
 
     if (type === 'external') {
@@ -232,7 +261,7 @@ export class MediaCardInternal extends Component<MediaCardProps, State> {
         resizeMode={resizeMode}
         isLazy={!isMobile}
         disableOverlay={disableOverlay}
-        useInlinePlayer={isMobile ? false : shouldPlayInline}
+        useInlinePlayer={isInlinePlayer}
         shouldOpenMediaViewer={shouldOpenMediaViewer}
         mediaViewerDataSource={{
           list: Array.from(mediaIdentifierMap.values()),