+++ packages/elements/emoji/example-helpers/demo-emoji-usage-components.tsx
@@ -1,9 +1,9 @@
 import { UsageClearEmojiResource } from '@atlaskit/util-data-test';
 import * as React from 'react';
 import { PureComponent } from 'react';
-import { EmojiProvider } from '../src/api/EmojiResource';
-import ResourcedEmoji from '../src/components/common/ResourcedEmoji';
-import { localStoragePrefix } from '../src/constants';
+import { EmojiProvider } from '../src/resource';
+import { ResourcedEmoji } from '../src/element';
+import { localStoragePrefix } from '../src/util/constants';
 import { EmojiDescription } from '../src/types';
 
 export interface EmojiUsageProps {