+++ packages/media/media-viewer/src/newgen/list.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Context, FileIdentifier } from '@atlaskit/media-core';
+import { Context, Identifier } from '@atlaskit/media-core';
 import { ItemViewer } from './item-viewer';
 import { MediaViewerFeatureFlags } from './domain';
 import { HeaderWrapper, hideControlsClassName, ListWrapper } from './styled';
@@ -9,16 +9,16 @@ import Header from './header';
 
 export type Props = Readonly<{
   onClose?: () => void;
-  onNavigationChange?: (selectedItem: FileIdentifier) => void;
+  onNavigationChange?: (selectedItem: Identifier) => void;
   showControls?: () => void;
   featureFlags?: MediaViewerFeatureFlags;
-  defaultSelectedItem: FileIdentifier;
-  items: FileIdentifier[];
+  defaultSelectedItem: Identifier;
+  items: Identifier[];
   context: Context;
 }>;
 
 export type State = {
-  selectedItem: FileIdentifier;
+  selectedItem: Identifier;
   previewCount: number;
 };
 
@@ -39,7 +39,7 @@ export class List extends React.Component<Props, State> {
     );
   }
 
-  renderContent(items: FileIdentifier[]) {
+  renderContent(items: Identifier[]) {
     const { context, onClose, featureFlags, showControls } = this.props;
     const { selectedItem } = this.state;
 
@@ -69,7 +69,7 @@ export class List extends React.Component<Props, State> {
     );
   }
 
-  onNavigationChange = (selectedItem: FileIdentifier) => {
+  onNavigationChange = (selectedItem: Identifier) => {
     const { onNavigationChange, showControls } = this.props;
     if (onNavigationChange) {
       onNavigationChange(selectedItem);