+++ packages/media/media-viewer/src/newgen/utils/index.ts
@@ -2,6 +2,8 @@ import {
   Context,
   isClientBasedAuth,
   FileIdentifier,
+  Identifier,
+  isFileIdentifier,
 } from '@atlaskit/media-core';
 import { stringify } from 'query-string';
 import { MediaCollectionItem } from '@atlaskit/media-store';
@@ -72,8 +74,18 @@ export const toIdentifier = (
 };
 
 export const getSelectedIndex = (
-  items: FileIdentifier[],
-  selectedItem: FileIdentifier,
+  items: Identifier[],
+  selectedItem: Identifier,
 ) => {
-  return items.findIndex(item => item.id === selectedItem.id);
+  return items.findIndex(item => {
+    if (isFileIdentifier(item) && isFileIdentifier(selectedItem)) {
+      return item.id === selectedItem.id;
+    }
+
+    if (!isFileIdentifier(item) && !isFileIdentifier(selectedItem)) {
+      return item.dataURI === selectedItem.dataURI;
+    }
+
+    return false;
+  });
 };