+++ packages/css-packs/css-reset/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/css-reset
 
+## 3.0.7
+- [patch] [6c0d9da30e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6c0d9da30e):
+
+  - Removes deprecated css property text-decoration-skip: ink. Users should no longer see a warning to update the property.
+
 ## 3.0.6
 - Updated dependencies [76299208e6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/76299208e6):
   - @atlaskit/docs@7.0.0